add shogun 2.0.0

+++ trunk/thirdsrc/shogun-1.1.0/src/shogun/lib/DataType.h
@@ -25,7 +25,7 @@ namespace shogun
 template<class T> class CCache;
 
 /** index */
-typedef int64_t index_t;
+typedef int32_t index_t;
 
 /** @brief shogun vector */
 template<class T> class SGVector
